remove post when deleted from user

+++ controllers/feed.js
@@ -149,6 +149,12 @@ exports.updatePost = (req, res, next) => {
                 throw error
             }
 
+            if (post.creator.toString() !== req.userId){
+                const error  = new Error('Not Authorized')
+                error.statusCode = 403
+                throw error
+            }
+
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -179,9 +185,21 @@ exports.deletePost = (req, res, next) => {
                 throw error
             }
 
+            if (post.creator.toString() !== req.userId){
+                const error  = new Error('Not Authorized')
+                error.statusCode = 403
+                throw error
+            }
             clearImage(post.imageUrl)
 
             return Post.findOneAndRemove(postId)
+        })
+        .then(result => {
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            user.posts.pull(postId)
+            return user.save()
         })
         .then(result => {
             res.status(200).json({message: 'Post Deleted'})
